Update super_clawbot
two controllers working, shoulder and arm able to be controlled by controller 1 also. maybe. claw doesn't seem to be working that way. Two sets of code to control to claw. One two control it with controller one and one to control the claw controller 2. Controller 2 battery is dead and not functioning to be able to check them at this time to see if the code works.